# shared/openapi.yaml
openapi: 3.1.0
info:
  title: Icarus API
  description: Contract v1 for the Icarus server (shared/api-contract.md). Times are RFC 3339 UTC with Z, except fields ending in _ms (epoch milliseconds). Metrics are estimates, not medical readings.
  version: 1.0.0
paths:
...
  /v1/alarm-dispatches:
    get:
      tags:
      - Alarms
      operationId: list_alarm_dispatches
      parameters:
      - name: limit
        in: query
        description: 1 to 200. Default 50.
        required: false
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: Newest first.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DispatchList'
        default:
          description: An error as application/problem+json (api-contract.md Errors).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      security:
      - session: []
      - device: []
  /v1/alarm-dispatches/{id}/ack:
    post:
      tags:
      - Alarms
      summary: |-
        The app reports what the phone and band did. The first ack sets `acked_at`; later acks replace
        the statuses.
      operationId: ack_alarm_dispatch
      parameters:
      - name: id
        in: path
        description: Id of the resource.
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AckBody'
        required: true
      responses:
        '204':
          description: Recorded. The first ack sets acked_at.
        default:
          description: An error as application/problem+json (api-contract.md Errors).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      security:
      - device: []
...
  /v1/alarms/pending:
    get:
      tags:
      - Alarms
      summary: Unacked dispatches from the last 10 minutes, for the app's pull (PLAN.md §9.3).
      operationId: pending_alarm_dispatches
      responses:
        '200':
          description: Unacked dispatches from the last 10 minutes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DispatchList'
        default:
          description: An error as application/problem+json (api-contract.md Errors).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      security:
      - device: []
...
components:
  schemas:
    AckBody:
      type: object
      required:
      - phone
      - band
      properties:
        band:
          type: string
        detail:
          type:
          - string
          - 'null'
        phone:
          type: string
      additionalProperties: false
...
    Dispatch:
      type: object
      description: '`Dispatch` in api-contract.md.'
      required:
      - id
      - created_at
      - attempts
      - status
      - rhythm
      properties:
        acked_at:
          type:
          - string
          - 'null'
          format: date-time
        alarm_id:
          type:
          - string
          - 'null'
          format: uuid
        attempts:
          type: integer
          format: int32
        band_status:
          type:
          - string
          - 'null'
        created_at:
          type: string
          format: date-time
        delivery_id:
          type:
          - string
          - 'null'
          format: uuid
        id:
          type: string
          format: uuid
        message:
          type:
          - string
          - 'null'
        phone_status:
          type:
          - string
          - 'null'
        rhythm:
          $ref: '#/components/schemas/Rhythm'
        status:
          $ref: '#/components/schemas/DispatchStatus'
...
    DispatchList:
      type: object
      required:
      - dispatches
      properties:
        dispatches:
          type: array
          items:
            $ref: '#/components/schemas/Dispatch'
    DispatchStatus:
      type: string
      enum:
      - pending
      - sent
      - acked
      - unacked
      - failed
...
    NamedRhythm:
      type: string
      description: Built-in rhythm names. PLAN.md §9.2.
      enum:
      - single
      - double
      - triple
      - long
      - ramp
      - sos
...
    Problem:
      type: object
      description: RFC 9457 problem document. Handlers build it in `error.rs`; this type only documents it.
      required:
      - type
      - title
      - status
      - detail
      properties:
        current:
          type:
          - object
          - 'null'
          description: The stored entity, on `conflict` responses only.
        detail:
          type: string
        status:
          type: integer
          format: int32
          minimum: 0
        title:
          type: string
        type:
          type: string
          description: '`urn:icarus:problem:<slug>`, where the slug is one of the contract''s list.'
...
    Rhythm:
      oneOf:
      - $ref: '#/components/schemas/NamedRhythm'
      - type: array
        items:
          $ref: '#/components/schemas/Step'
      description: A built-in name, or an explicit list of steps.
...
    Step:
      oneOf:
      - type: object
        required:
        - preset
        - loops
        - type
        properties:
          loops:
            type: integer
            format: int32
            minimum: 0
          preset:
            type: integer
            format: int32
            minimum: 0
          type:
            type: string
            enum:
            - buzz
      - type: object
        required:
        - ms
        - type
        properties:
          ms:
            type: integer
            format: int32
            minimum: 0
          type:
            type: string
            enum:
            - pause
      description: One rhythm step. A buzz loop counts as 1 s toward the 30 s limit.
...
  securitySchemes:
    device:
      type: http
      scheme: bearer
    session:
      type: apiKey
      in: cookie
      name: icarus_session
      description: 'Website session from POST /v1/auth/login. Non-GET requests also send X-Icarus-CSRF: 1.'
...